Separando callbacks do main_menu
