+++ src/controller/call_backs_main_menu.h
@@ -1,0 +1,9 @@
+#ifndef CALL_BACKS_MAIN_MENU_H
+
+#define CALL_BACKS_MAIN_MENU_H
+
+#include <gtk/gtk.h>
+
+void execute( GtkWidget *widget, gpointer data );
+
+#endif
