+++ src/controller/controller.c
@@ -64,23 +64,6 @@ int controller( int argc, char *argv[] ) {
     return status;
 }
 
-// CALLBACKs of main_menu
-void execute( GtkWidget *widget, gpointer data ) {
-    int number_of_fields = 0;
-    char **argv;
-    argv =
-        (char **)malloc( sizeof( char ** ) * ( 2 * NUMBER_OF_FIELDS + 1 + ARGS_TO_OUTPUT_GOURCE ) );
-    if( argv != NULL ) {
-        add_to_argv_valid_field( &gource_settings, argv, &number_of_fields );
-        call_prog( "/usr/bin/gource", argv, number_of_fields );
-    } else {
-        free_memory( &gource_settings, true );
-    }
-    print_gource( &gource_settings );
-}
-
-// CALLBACKs of video_page
-
 // CALLBACKs of subtitle_page
 
 void set_subtitle_file( GtkWidget *widget, gpointer data ) {
